The modality selection is broken.
The slide loader is working but currently the modality selection is broken. I am trying to make it more general using the package "python-magic".
