+++ taplt/mediaViewingWidgets/examples/slide_example.py
@@ -1,6 +1,5 @@
 from PyQt6.QtWidgets import *
-from media_viewing_widgets import SlideView
-from graphics_view import GraphicsView
+from taplt.mediaViewingWidgets.slide_view import SlideView
 
 
 if __name__ == '__main__':
@@ -12,7 +11,7 @@ if __name__ == '__main__':
     scene = QGraphicsScene()
     scene.addItem(slide_view)
 
-    viewer = GraphicsView(scene)
+    viewer = SlideView(scene)
     viewer.resize(1000, 600)
     viewer.show()
     slide_view.load_new_image(QFileDialog().getOpenFileName()[0])
