+++ taplt/utils/project_structure.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 import filetype
+import magic
 
 
 class Structure:
@@ -34,18 +35,14 @@ def create_project_structure(project_path: str):
 def modality(filepath: str):
     """This method uses the 'filetype' library to detect the type of a given file
     returns: 0 if video, 1 if image, 2 if whole slide image, -1 if none of the above"""
-    detection = filetype.guess(filepath)
+    mime = magic.Magic(mime=True)
+    detection = mime.from_file(filepath)
 
-    if detection is None:
-        if filepath.endswith(".tif"):
-            return 2
-        else:
-            RuntimeError("Unknown Filetype!")
-
-    detection = detection.mime
     if detection.startswith('video'):
         return 0
     elif detection.startswith('image'):
+        if detection.endswith('tiff'):
+            return 2
         return 1
     else:
         return -1
